Client can send checks to server.

+++ client/launch.py
@@ -10,8 +10,6 @@ from settings import get_settings
 
 
 def launch_smt4_client(*args: Sequence[str]) -> None:
-    from .client import main
-
     parser = get_base_parser()
     parser.add_argument("patch_file", default="", type=str, nargs="?", help="Path to an Archipelago patch file")
     parser.add_argument("--name", default=None, help="Slot Name to connect as.")
@@ -23,13 +21,16 @@ def launch_smt4_client(*args: Sequence[str]) -> None:
     if patch_file != "":
         metadata, output_file = Patch.create_rom_file(patch_file)
 
-    azahar_path = get_settings().smt4.azahar_path
-    azahar_process = subprocess.Popen([azahar_path, output_file])
-
-    colorama.just_fix_windows_console()
-
+    azahar_process = None
     try:
+        azahar_path = get_settings().smt4.azahar_path
+        azahar_process = subprocess.Popen([azahar_path, output_file])
+
+        colorama.just_fix_windows_console()
+
+        from .client import main
         asyncio.run(main(launch_args))
     finally:
-        azahar_process.kill()
+        if azahar_process is not None:
+            azahar_process.kill()
         colorama.deinit()
